@@ -7,7 +7,7 @@ name = "assetutilities"
 version = "0.1.0"
 description = "Standardized project configuration for assetutilities"
 readme = "README.md"
-requires-python = ">=3.8"
+requires-python = ">=3.9"
 license = {text = "MIT"}
 authors = [
     {name = "Development Team", email = "dev@example.com"}
